Refactor move code decrypt file from lib CryptoAES to local extension
