+++ src/pt/randomscan/build.gradle
@@ -9,6 +9,5 @@ apply from: "$rootDir/common.gradle"
 
 dependencies {
     implementation project(':lib:randomua')
-    implementation project(':lib:cryptoaes')
     implementation project(':lib:zipinterceptor')
 }
