+++ src/pt/randomscan/src/eu/kanade/tachiyomi/extension/pt/randomscan/LuraZipInterceptor.kt
@@ -1,14 +1,34 @@
 package eu.kanade.tachiyomi.extension.pt.randomscan
 
-import eu.kanade.tachiyomi.lib.cryptoaes.CryptoAES
 import eu.kanade.tachiyomi.lib.zipinterceptor.ZipInterceptor
 import okhttp3.Request
 import okhttp3.Response
 import java.io.ByteArrayInputStream
 import java.io.InputStream
 import java.nio.charset.StandardCharsets
+import java.security.MessageDigest
+import javax.crypto.Cipher
+import javax.crypto.SecretKey
+import javax.crypto.spec.IvParameterSpec
+import javax.crypto.spec.SecretKeySpec
 
 class LuraZipInterceptor : ZipInterceptor() {
+    fun decryptFile(encryptedData: ByteArray, keyBytes: ByteArray): ByteArray {
+        val keyHash = MessageDigest.getInstance("SHA-256").digest(keyBytes)
+
+        val key: SecretKey = SecretKeySpec(keyHash, "AES")
+
+        val counter = encryptedData.copyOfRange(0, 8)
+        val iv = IvParameterSpec(counter)
+
+        val cipher = Cipher.getInstance("AES/CTR/NoPadding")
+        cipher.init(Cipher.DECRYPT_MODE, key, iv)
+
+        val decryptedData = cipher.doFinal(encryptedData.copyOfRange(8, encryptedData.size))
+
+        return decryptedData
+    }
+
     override fun requestIsZipImage(request: Request): Boolean {
         return request.url.pathSegments.contains("cap-download")
     }
@@ -19,7 +39,7 @@ class LuraZipInterceptor : ZipInterceptor() {
         }.toByteArray(StandardCharsets.UTF_8)
         val encryptedData = response.body.bytes()
 
-        val decryptedData = CryptoAES.decryptFile(encryptedData, keyData, "AES/CTR/NoPadding")
+        val decryptedData = decryptFile(encryptedData, keyData)
         return ByteArrayInputStream(decryptedData)
     }
 }
